% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_exams2form.R
\name{n_exams2form}
\alias{n_exams2form}
\title{Generate Text for Quarto with Exams2webexercises}
\usage{
n_exams2form(file_path, n)
}
\arguments{
\item{file_path}{.}

\item{n}{An integer specifying the number of questions to generate.}
}
\value{
A character vector of lines to be used within Quarto.
}
\description{
This function generates text for Quarto with Exams2webexercises.
}
\examples{
\dontrun{
n_exams2form(c("example_exam1.Rmd", "example_exam2.Rmd"), 5)
}
}
